refactor: 앨범 다운로드 hook 분리
- username 찾는 로직 hook으로 분리
